==> notebooks/0_data_explorer.ipynb <==
# Data Explorer

## Emotion Face Classifier Notebook 0

Designed to show loaded data used throughout each step.

Variable names for each dictionary and dataframe should be consistent across notebooks.

**NOTE:** Some files created from running notebooks and can not be explored until generated.

%load_ext autoreload
%autoreload 2

from datascifuncs import check_directory_name, load_json, print_json

main_dir = 'EmotionFaceClassifier'
check_directory_name(main_dir)

from utils.data_loader import DataLoader

### Data Loader

Loads common json and csv files. By default, all following will be loaded. Each item can be set to False to skip. None is returned if file does not exist.

settings.json also sets each primary key/value pair as attribute of the class instance.

Key JSON files, stored in configs:
- settings.json
- unsupervised_models.json
- vectorized_models.json

Key csv files, stored in data:
- fer2013.csv
- der2013_processed.csv

# Initialize DataLoader with boolean flags
data_loader = DataLoader(
    settings=True, 
    unsupervised=True, 
    vectorized=True, 
    initial_data=True, 
    processed_data=True
)

# Call method to read in data
data_loader.load_all_data()

for attr in data_loader.__dict__.keys():
    print(f"{attr}")


# print_json(common_dicts)


# # Load all data


# # Access the loaded data directly as attributes
# print(data_loader.initial_csv.head() if data_loader.initial_csv is not None else "Initial CSV data not loaded")
# print(data_loader.processed_csv.head() if data_loader.processed_csv is not None else "Processed CSV data not loaded")
# print(data_loader.settings_json if data_loader.settings_json is not None else "Settings JSON not loaded")
# print(data_loader.unsupervised_json if data_loader.unsupervised_json is not None else "Unsupervised JSON not loaded")
# print(data_loader.vectorized_json if data_loader.vectorized_json is not None else "Vectorized JSON not loaded")
