==> notebooks/0_reset_project.ipynb <==
# Reset Project (Optional)

## Emotion Face Classifier Notebook 0

Removes existing images and intermediate files for a fresh analysis.

#### NOTE: This notebook completely removes data including images, models, and results. If you are not sure, do NOT use this notebook. Remove unwanted files by hand if needed.

from os.path import join
from datascifuncs.tidbit_tools import check_directory_name
from datascifuncs.reset_project import remove_files, remove_directories, reset_notebooks

# Ensure working directory for correct filepaths
main_dir = 'EmotionFaceClassifier'
check_directory_name(main_dir)

### Remove Filepaths
Set a list of file patterns to be matched using glob.

remove_files will remove all files matching any of the patterns.

remove_files(file_patterns=['data/*_paths.csv'])

### Remove Directories

Create a list of directories to be removed.

Each of those directories will be removed entirely. 

img_dir = 'imgs'
data_dir = 'data'
model_dir = 'models'

# Directory with pivot and summary data tables
intermediate_data = join(data_dir, 'intermediate')

# Directories with image files extracted from initial csv file
training_data = join(data_dir, 'Training')
testing_data = join(data_dir, 'Testing')

remove_list =[img_dir, intermediate_data, training_data, testing_data, model_dir]

remove_directories(directories=remove_list)

### Reset Jupyter Notebooks
Reset notebooks in the specified directory ("notebooks").

Exclude current notebook to avoid issues due to modifying an open file.

    Note: excluded_notebooks accepts a list so that additional notebooks can be skipped.

excluded_notebooks = ['notebooks/0_reset_project.ipynb']
reset_notebooks(notebook_dir='notebooks', excluded_notebooks=excluded_notebooks)
