==> notebooks/3_eda_image_properties.ipynb <==
# Exploratory Data Analysis (EDA) Image Properties 

## Emotion Face Classifier Notebook 3

Creates multiple visualizations of images by emotional category including:
- Sample images
- Pixel intensity histograms
- Composite images (average and median)

%load_ext autoreload
%autoreload 2

import os
import itertools
import pandas as pd

from datascifuncs.tidbit_tools import load_json, write_json, print_json, check_directory_name

main_dir = 'EmotionFaceClassifier'
check_directory_name(main_dir)

from utils.preprocessing import (
    preprocess_images,
    plot_face_matrix
)

from utils.image_properties import (
    generate_sample_images,
    generate_pixel_intensities,
    generate_composite_faces
)

### Load and split data

Read in data from csv.

Select only the training portion of the data.

Load in project configuration file.

# Read in FER 2013 data
fer2013_path = 'data/fer2013_paths.csv'
fer2013 = pd.read_csv(fer2013_path)

# Check data
fer2013.head()

# Select training data
# Ensure number of rows drops appropriately
print(fer2013.shape)
train_df = fer2013[fer2013['usage']=='Training']
print(train_df.shape)

# Load common dicts from json config file
common_dicts = load_json('./configs/input_mappings.json')

# Get subset of emo-color mappings
color_dict = common_dicts['color_dict']
color_dict

image_dir = os.path.join('images', 'comparisons')
os.makedirs(image_dir, exist_ok=True)

### Select and Display Sample Images

Randomly sample images from each category.

Set output path to save image. 

# Create dictionary of n images for each category
emo_samples = generate_sample_images(train_df, n=5, cat_col='emotion', path_col='img_path')

# Set path to save sample matrix to
sample_imgs_save_path = os.path.join(image_dir, 'sample_images.png')

# Plot sample images
plot_face_matrix(
    image_dict=emo_samples, 
    row_labels=None,
    group_colors=color_dict, 
    save_path=sample_imgs_save_path,
    main_title='Sample Images'
)

### Visualize Pixel Intensities

Convert images into matrix for analysis. 

Set output path.

Save image to path.

# Convert data into X and y arrays
X_train, y_train = preprocess_images(fer2013, usage='Training')

# Set path for output
pixel_imgs_save_path = os.path.join(image_dir, 'pixel_intensities.png')

# Calculate intensities and save to path
generate_pixel_intensities(
    X_train, 
    y_train, 
    color_dict=color_dict, 
    save_path=pixel_imgs_save_path
)

### Composite Faces

Use mean and median calculations on categories for composite category representations.

Plot and save images to file path

# Get dictionary of categories and composite images
# List of row labels for axis labels when plotting
composite_face_dict, row_labels = generate_composite_faces(X_train, y_train, overall=True)

# Set output path
composite_imgs_save_path = os.path.join(image_dir, 'composite_faces.png')

# Create visual and save to path
plot_face_matrix(
    image_dict=composite_face_dict, 
    row_labels=row_labels,
    group_colors=color_dict, 
    save_path=composite_imgs_save_path,
    main_title='Composite Faces'
)
