==> notebooks/4_unsupervised_feature_extraction.ipynb <==
# Unsupervised Feature Extraction

## Emotion Face Classifier Notebook 4

Load model parameters from unsupervised_models.json.

Iterative through model options for feature extraction.

Outputs from each run saves to dir based on model settings-existing dirs will be skipped.

%load_ext autoreload
%autoreload 2

import os
import pandas as pd

from datascifuncs.tidbit_tools import load_json, print_json, check_directory_name

main_dir = 'EmotionFaceClassifier'
check_directory_name(main_dir)

from utils.preprocessing import preprocess_images
from utils.feature_extraction import generate_analysis_paths, run_single_analysis

### Load Configs and Data

Read in json with model parameters.

Read in data and select training data for analysis.

# Read in FER 2013 data
fer2013_path = 'data/fer2013_paths.csv'
fer2013 = pd.read_csv(fer2013_path)

fer2013.head()

# Load common dicts from json config file
common_dicts = load_json('./configs/input_mappings.json')

# Get subset of emo-color mappings
color_dict = common_dicts['color_dict']
color_dict

X_train, y_train = preprocess_images(fer2013, usage='Training', flatten=True)

model_dict = load_json('configs/unsupervised_models.json')

model_dict

### Run Analyses

For each model, skip if results exist or run analysis.

Model details saved for each iteration. 

for model_name, model_details in model_dict.items():
    analysis_details = generate_analysis_paths(model_details)
    analysis_details['color_map'] = color_dict
    run_single_analysis(X_train, y_train, analysis_details)

print('Unsupervised feature extraction complete!')
